Title: 삼 오 무, Time: 86ms, Memory: 8MB, Status: Passed - CodeTree

+++ 241023/삼 오 무/three-five-moo.java
@@ -33,8 +33,8 @@ public class Main {
     
     }
 
-    private static int cal(int number) {
-        int sum = 0;
+    private static long cal(long number) {
+        long sum = 0;
 
         sum = sum + number/3;
         sum = sum + number/5;
